requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-dotenv==1.1.1
GitPython==3.1.34
haystack-ai==2.16.0
pgvector-haystack
sentence-transformers>=5.0.0,<6.0.0
transformers>=4.54.1,<5.0.0
psycopg[binary]
asyncpg
passlib
pyjwt
pydantic-settings>=0.2.2
pydantic[email]
tree-sitter-languages
tree-sitter
jose
aiofiles